% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textaObject.R
\name{texta}
\alias{texta}
\title{The \code{texta} object}
\description{
The \code{texta} object exposes formatted results, the REST API JSON
response, and the HTTP request:

\itemize{
  \item \code{result} the results in \code{data.frame} format
  \item \code{json} the REST API JSON response
  \item \code{request} the HTTP request
}
}
\author{
Phil Ferriere \email{pferriere@hotmail.com}
}
\seealso{
Other res: \code{\link{textatopics}}
}

