% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textatopicsObject.R
\name{textatopics}
\alias{textatopics}
\title{The \code{textatopics} object}
\description{
The \code{textatopics} object exposes formatted results for the
\code{\link{textaDetectTopics}} API, this REST API's JSON response, and the
HTTP request:

\itemize{
  \item \code{status} the operation's current status ("NotStarted"|"Running"|"Succeeded"|"Failed")
  \item \code{documents} a \code{data.frame} with the documents and a unique
  string ID for each
  \item \code{topics} a \code{data.frame} with the identified topics, a
  unique string ID for each, and a prevalence score for each topic (count of
  documents assigned to topic)
  \item \code{topicAssignments} a \code{data.frame} with all the topics
  (identified by their topic ID) assigned to each document (identified by
  their document ID), and a distance score for each topic assignment (between
  0 and 1; the lower the distance score the stronger the topic affiliation)
  \item \code{json} the REST API JSON response
  \item \code{request} the HTTP request
}
}
\author{
Phil Ferriere \email{pferriere@hotmail.com}
}
\seealso{
Other res: \code{\link{texta}}
}

